backend/pyproject.toml
[tool.poetry]
name = "leaders-2021-05"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
psycopg2 = "^2.9.1"
djangorestframework = "^3.12.4"
python-dotenv = "^0.19.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
